refactor: split EventLiveSeriesChart into focused sub-components
Extracts 8 files from the 1745 LOC monolith (now 582 LOC):
- eventLiveSeriesChartUtils.ts (pure helpers + constants)
- AnimatedCountdownValue.tsx (countdown digit animations)
- EventLiveSeriesChartHeader.tsx (price header + countdown tooltip)
- EventLiveSeriesChartOverlay.tsx (target line + price guide)
- EventLiveSeriesViewSwitch.tsx (live/market toggle)
- useLiveSeriesClock.ts (RAF clock hook)
- useLiveSeriesPriceSnapshot.ts (price snapshot fetch hook)
- useLiveSeriesWebSocket.ts (WS connection + reconnect hook)
