+++ src/app/[locale]/(platform)/event/[slug]/_components/EventLiveSeriesChart.tsx
@@ -2,675 +2,58 @@
 
 import type { Event, EventLiveChartConfig, EventSeriesEntry } from '@/types'
 import type { DataPoint, PredictionChartProps, SeriesConfig } from '@/types/PredictionChartTypes'
-import { ChartLineIcon, ChevronsDownIcon, ChevronsUpIcon, TriangleIcon } from 'lucide-react'
 import dynamic from 'next/dynamic'
-import { useEffect, useMemo, useState } from 'react'
-import SiteLogoIcon from '@/components/SiteLogoIcon'
-import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip'
+import { useMemo, useState } from 'react'
 import { useSiteIdentity } from '@/hooks/useSiteIdentity'
 import { useWindowSize } from '@/hooks/useWindowSize'
-import { formatCurrency } from '@/lib/formatters'
 import { cn } from '@/lib/utils'
-import { createWebSocketReconnectController } from '@/lib/websocket-reconnect'
+import { useLiveSeriesClock } from '../_hooks/useLiveSeriesClock'
+import { useLiveSeriesPriceSnapshot } from '../_hooks/useLiveSeriesPriceSnapshot'
+import { useLiveSeriesWebSocket } from '../_hooks/useLiveSeriesWebSocket'
+import {
+  buildAxis,
+  formatDateAtTimezone,
+  formatTimeAtTimezone,
+  formatUsd,
+  getVisibleCountdownUnits,
+  hexToRgba,
+  inferIntervalMsFromSeriesSlug,
+  isUsEquityMarketOpen,
+  LIVE_CHART_HEIGHT,
+  LIVE_CHART_MARGIN_BOTTOM,
+  LIVE_CHART_MARGIN_LEFT,
+  LIVE_CHART_MARGIN_RIGHT,
+  LIVE_CHART_MARGIN_TOP,
+  LIVE_CURRENT_MARKER_OFFSET_X,
+  LIVE_CURSOR_GUIDE_TOP,
+  LIVE_PLOT_CLIP_RIGHT_PADDING,
+  LIVE_TARGET_MAX_BOTTOM_OFFSET,
+  LIVE_WINDOW_MS,
+  LIVE_X_AXIS_LEFT_LABEL_GUARD_MS,
+  LIVE_X_AXIS_STEP_MS,
+  MAX_POINTS,
+  normalizeLiveChartPrice,
+  normalizeSubscriptionSymbol,
+  parseUtcDate,
+  readPersistedLivePrice,
+  resolveEventEndTimestamp,
+  SERIES_KEY,
+  toCountdownLeftLabel,
+} from '../_utils/eventLiveSeriesChartUtils'
 import {
   resolveLiveSeriesDeltaDisplayDigits,
   resolveLiveSeriesPriceDisplayDigits,
-  resolveLiveSeriesTopicPriceDigits,
 } from '../_utils/liveSeriesPricePrecision'
 import EventChart from './EventChart'
+import EventLiveSeriesChartHeader from './EventLiveSeriesChartHeader'
+import EventLiveSeriesChartOverlay from './EventLiveSeriesChartOverlay'
+import EventLiveSeriesViewSwitch from './EventLiveSeriesViewSwitch'
 import EventSeriesPills from './EventSeriesPills'
 
 const PredictionChart = dynamic<PredictionChartProps>(
   () => import('@/components/PredictionChart'),
   { ssr: false, loading: () => <div className="h-83 w-full" /> },
 )
-
-const SERIES_KEY = 'live_price'
-const LIVE_WINDOW_MS = 40 * 1000
-const LIVE_HISTORY_BUFFER_MS = 8 * 1000
-const LIVE_DATA_RETENTION_MS = LIVE_WINDOW_MS + LIVE_HISTORY_BUFFER_MS
-const LIVE_CLOCK_FRAME_MS = 1000 / 30
-const LIVE_X_AXIS_STEP_MS = 10 * 1000
-const LIVE_X_AXIS_LEFT_LABEL_GUARD_MS = 3600
-const LIVE_MAX_Y_AXIS_TICKS = 6
-const MAX_POINTS = 4000
-const LIVE_WS_USE_ONLY_LAST_UPDATE_PER_MESSAGE = true
-const LIVE_CHART_HEIGHT = 332
-const LIVE_CHART_MARGIN_TOP = 22
-const LIVE_CHART_MARGIN_BOTTOM = 52
-const LIVE_CHART_MARGIN_RIGHT = 40
-const LIVE_CHART_MARGIN_LEFT = 0
-const LIVE_CURSOR_GUIDE_TOP = 10
-const LIVE_TARGET_MAX_BOTTOM_OFFSET = 10
-const LIVE_CURRENT_MARKER_OFFSET_X = 0
-const LIVE_PLOT_CLIP_RIGHT_PADDING = 22
-const LIVE_PRICE_STORAGE_PREFIX = 'kuest-live-last-price'
-
-interface PersistedLivePrice {
-  price: number
-  timestamp: number
-}
-
-interface LiveSeriesPriceSnapshot {
-  series_slug: string
-  instrument: string
-  interval: '5m' | '15m' | '1h' | '4h' | '1d'
-  source: 'chainlink' | 'massive'
-  interval_ms: number
-  event_window_start_ms: number
-  event_window_end_ms: number
-  opening_price: number | null
-  closing_price: number | null
-  latest_price: number | null
-  latest_window_end_ms: number | null
-  latest_source_timestamp_ms: number | null
-  is_event_closed: boolean
-}
-
-function normalizeTimestamp(value: unknown, fallbackTimestamp = 0) {
-  const numeric = typeof value === 'number' ? value : Number(value)
-  if (!Number.isFinite(numeric)) {
-    return fallbackTimestamp
-  }
-  return numeric < 1e12 ? numeric * 1000 : numeric
-}
-
-function buildLivePriceStorageKey(topic: string, symbol: string) {
-  return `${LIVE_PRICE_STORAGE_PREFIX}:${topic.trim().toLowerCase()}:${symbol.trim().toUpperCase()}`
-}
-
-function readPersistedLivePrice(topic: string, symbol: string): PersistedLivePrice | null {
-  if (typeof window === 'undefined') {
-    return null
-  }
-
-  try {
-    const key = buildLivePriceStorageKey(topic, symbol)
-    const raw = window.localStorage.getItem(key)
-    if (!raw) {
-      return null
-    }
-
-    const parsed = JSON.parse(raw) as Partial<PersistedLivePrice>
-    const price = Number(parsed.price)
-    const timestamp = normalizeTimestamp(parsed.timestamp)
-
-    if (!Number.isFinite(price) || price <= 0) {
-      return null
-    }
-
-    return {
-      price,
-      timestamp,
-    }
-  }
-  catch {
-    return null
-  }
-}
-
-function writePersistedLivePrice(topic: string, symbol: string, price: number, timestamp: number) {
-  if (typeof window === 'undefined' || !Number.isFinite(price) || price <= 0) {
-    return
-  }
-
-  try {
-    const key = buildLivePriceStorageKey(topic, symbol)
-    const payload: PersistedLivePrice = {
-      price,
-      timestamp: normalizeTimestamp(timestamp),
-    }
-    window.localStorage.setItem(key, JSON.stringify(payload))
-  }
-  catch {
-  }
-}
-
-function matchesSymbol(symbol: string | null, targetSymbol: string) {
-  if (!targetSymbol) {
-    return true
-  }
-  if (!symbol) {
-    return false
-  }
-  return symbolsAreEquivalent(symbol, targetSymbol)
-}
-
-function extractPointsFromArray(
-  entries: any[],
-  fallbackSymbol: string | null = null,
-  fallbackTimestamp = 0,
-) {
-  if (!Array.isArray(entries) || entries.length === 0) {
-    return []
-  }
-
-  const points: Array<{ price: number, timestamp: number, symbol: string | null }> = []
-
-  for (const point of entries) {
-    if (!point || typeof point !== 'object') {
-      continue
-    }
-
-    const price = Number(point.value ?? point.price ?? point.p)
-    if (!Number.isFinite(price) || price <= 0) {
-      continue
-    }
-
-    const rawSymbol = point.symbol ?? point.pair ?? point.asset ?? point.base ?? fallbackSymbol
-    const symbol = typeof rawSymbol === 'string' ? rawSymbol : null
-    const timestamp = normalizeTimestamp(point.timestamp ?? point.ts ?? point.t, fallbackTimestamp)
-
-    points.push({ price, timestamp, symbol })
-  }
-
-  return points
-}
-
-function extractLivePriceUpdates(payload: any, topic: string, symbol: string, fallbackTimestamp = 0) {
-  if (!payload || typeof payload !== 'object') {
-    return []
-  }
-
-  const updates: Array<{ price: number, timestamp: number, symbol: string | null }> = []
-  const candidates: any[] = []
-  if (Array.isArray(payload)) {
-    candidates.push(...payload)
-  }
-  else {
-    candidates.push(payload)
-  }
-
-  if (payload?.payload && typeof payload.payload === 'object') {
-    candidates.push(payload.payload)
-  }
-
-  if (Array.isArray(payload?.data)) {
-    candidates.push(...payload.data)
-  }
-  else if (payload?.data && typeof payload.data === 'object') {
-    candidates.push(payload.data)
-  }
-
-  for (const candidate of candidates) {
-    if (!candidate || typeof candidate !== 'object') {
-      continue
-    }
-
-    const candidateTopic = candidate.topic ?? candidate?.data?.topic ?? candidate?.payload?.topic ?? candidate?.stream
-    if (typeof candidateTopic === 'string' && candidateTopic !== topic) {
-      continue
-    }
-
-    const rawSymbol = candidate?.data?.symbol
-      ?? candidate?.symbol
-      ?? candidate?.data?.pair
-      ?? candidate?.pair
-      ?? candidate?.data?.asset
-      ?? candidate?.asset
-      ?? candidate?.data?.base
-      ?? candidate?.base
-      ?? candidate?.payload?.symbol
-
-    const candidateSymbol = typeof rawSymbol === 'string' ? rawSymbol : null
-
-    if (Array.isArray(candidate?.data)) {
-      updates.push(...extractPointsFromArray(candidate.data, candidateSymbol, fallbackTimestamp))
-    }
-
-    if (Array.isArray(candidate?.payload?.data)) {
-      updates.push(...extractPointsFromArray(candidate.payload.data, candidateSymbol, fallbackTimestamp))
-    }
-
-    const rawPrice = candidate?.data?.price
-      ?? candidate?.price
-      ?? candidate?.data?.value
-      ?? candidate?.value
-      ?? candidate?.data?.p
-      ?? candidate?.p
-      ?? candidate?.payload?.value
-      ?? candidate?.payload?.price
-
-    const price = Number(rawPrice)
-    if (!Number.isFinite(price) || price <= 0) {
-      continue
-    }
-
-    const timestamp = normalizeTimestamp(
-      candidate?.data?.timestamp
-      ?? candidate?.timestamp
-      ?? candidate?.data?.ts
-      ?? candidate?.ts
-      ?? candidate?.data?.t
-      ?? candidate?.t
-      ?? candidate?.payload?.timestamp,
-      fallbackTimestamp,
-    )
-
-    updates.push({
-      price,
-      timestamp,
-      symbol: candidateSymbol,
-    })
-  }
-
-  const filtered = updates.filter(update => !update.symbol || matchesSymbol(update.symbol, symbol))
-  if (!filtered.length) {
-    return []
-  }
-
-  const sorted = filtered.sort((a, b) => a.timestamp - b.timestamp)
-  const deduped: Array<{ price: number, timestamp: number, symbol: string | null }> = []
-
-  for (const update of sorted) {
-    const last = deduped.at(-1)
-    if (last && last.timestamp === update.timestamp) {
-      deduped[deduped.length - 1] = update
-      continue
-    }
-    deduped.push(update)
-  }
-
-  return deduped
-}
-
-function isSnapshotMessage(payload: any) {
-  if (!payload || typeof payload !== 'object') {
-    return false
-  }
-
-  const messageType = String(payload?.type ?? '').trim().toLowerCase()
-  if (messageType !== 'subscribe') {
-    return false
-  }
-
-  return Array.isArray(payload?.payload?.data) || Array.isArray(payload?.data)
-}
-
-function buildAxis(values: number[], fractionDigits = 2) {
-  const resolvedFractionDigits = Math.max(0, Math.min(6, Math.floor(fractionDigits)))
-  const visibleStep = 1 / 10 ** resolvedFractionDigits
-  function roundAxisValue(value: number) {
-    return Number(value.toFixed(resolvedFractionDigits))
-  }
-
-  if (!values.length) {
-    return { min: 0, max: 1, ticks: [0, 1] }
-  }
-
-  const minValue = Math.min(...values)
-  const maxValue = Math.max(...values)
-  const midpoint = (minValue + maxValue) / 2
-
-  if (maxValue - minValue < visibleStep / 2 && Math.abs(midpoint) >= 50) {
-    const center = roundAxisValue(midpoint)
-    const axisMin = roundAxisValue(center - visibleStep)
-    const axisMax = roundAxisValue(center + visibleStep)
-    return { min: axisMin, max: axisMax, ticks: [axisMin, center, axisMax] }
-  }
-
-  const minSpan = Math.max(Math.abs(midpoint) * 0.00002, Math.abs(midpoint) >= 1 ? 0.002 : 0.0002)
-  const span = Math.max(minSpan, maxValue - minValue)
-  const padding = Math.max(span * 0.08, minSpan * 0.08)
-  const rawMin = minValue - padding
-  const rawMax = maxValue + padding
-
-  const targetTicks = 4
-  const rawStep = (rawMax - rawMin) / Math.max(1, targetTicks - 1)
-  const magnitude = 10 ** Math.floor(Math.log10(rawStep))
-  const stepRatio = rawStep / magnitude
-  const stepMultiplier = stepRatio >= 5 ? 5 : stepRatio >= 2 ? 2 : 1
-  const initialStep = Math.max(stepMultiplier * magnitude, visibleStep)
-
-  function nextNiceStep(currentStep: number) {
-    const currentMagnitude = 10 ** Math.floor(Math.log10(currentStep))
-    const normalized = currentStep / currentMagnitude
-
-    if (normalized < 2) {
-      return 2 * currentMagnitude
-    }
-
-    if (normalized < 5) {
-      return 5 * currentMagnitude
-    }
-
-    return 10 * currentMagnitude
-  }
-
-  function buildTicksForStep(step: number) {
-    const axisMin = Math.floor(rawMin / step) * step
-    const axisMax = Math.ceil(rawMax / step) * step
-    const ticks: number[] = []
-
-    for (let value = axisMin; value <= axisMax + step * 1e-6; value += step) {
-      ticks.push(value)
-    }
-
-    return { axisMin, axisMax, ticks }
-  }
-
-  let step = initialStep
-  let axis = buildTicksForStep(step)
-  let attempts = 0
-
-  while (axis.ticks.length > LIVE_MAX_Y_AXIS_TICKS && attempts < 8) {
-    step = nextNiceStep(step)
-    axis = buildTicksForStep(step)
-    attempts += 1
-  }
-
-  const ticks: number[] = []
-  const seenTicks = new Set<number>()
-
-  for (const value of axis.ticks) {
-    const roundedValue = roundAxisValue(value)
-    if (seenTicks.has(roundedValue)) {
-      continue
-    }
-
-    seenTicks.add(roundedValue)
-    ticks.push(roundedValue)
-  }
-
-  return {
-    min: roundAxisValue(axis.axisMin),
-    max: roundAxisValue(axis.axisMax),
-    ticks,
-  }
-}
-
-function keepWithinLiveWindow(points: DataPoint[], cutoffMs: number) {
-  if (!points.length) {
-    return points
-  }
-
-  const trimmed = points.filter(point => point.date.getTime() >= cutoffMs)
-  if (trimmed.length > 0) {
-    return trimmed
-  }
-
-  const lastPoint = points.at(-1)
-  const lastPrice = lastPoint?.[SERIES_KEY]
-  if (typeof lastPrice !== 'number' || !Number.isFinite(lastPrice)) {
-    return []
-  }
-
-  return [{
-    date: new Date(cutoffMs + 1),
-    [SERIES_KEY]: lastPrice,
-  }]
-}
-
-function formatUsd(value: number, digits = 2) {
-  return formatCurrency(value, {
-    minimumFractionDigits: digits,
-    maximumFractionDigits: digits,
-  })
-}
-
-function normalizeLiveChartPrice(price: number, topic: string) {
-  if (!Number.isFinite(price) || price <= 0) {
-    return null
-  }
-
-  const digits = resolveLiveSeriesTopicPriceDigits(topic)
-  const factor = 10 ** digits
-  return Math.round(price * factor) / factor
-}
-
-function normalizeSubscriptionSymbol(topic: string, symbol: string) {
-  const trimmed = symbol.trim()
-  if (!trimmed) {
-    return trimmed
-  }
-
-  if (topic.trim().toLowerCase() === 'equity_prices') {
-    return trimmed.split(/[/-]/)[0]?.trim().toUpperCase() || trimmed.toUpperCase()
-  }
-
-  return trimmed.toLowerCase()
-}
-
-function normalizeComparableSymbol(symbol: string) {
-  return symbol.toLowerCase().replace(/[^a-z0-9]/g, '')
-}
-
-function symbolsAreEquivalent(symbol: string, target: string) {
-  const normalizedSymbol = normalizeComparableSymbol(symbol)
-  const normalizedTarget = normalizeComparableSymbol(target)
-  if (!normalizedSymbol || !normalizedTarget) {
-    return false
-  }
-
-  if (normalizedSymbol === normalizedTarget) {
-    return true
-  }
-
-  const symbolNoQuote = normalizedSymbol.replace(/(usd|usdt)$/i, '')
-  const targetNoQuote = normalizedTarget.replace(/(usd|usdt)$/i, '')
-  return symbolNoQuote === targetNoQuote
-}
-
-function hexToRgba(color: string, alpha: number) {
-  const hex = color.trim().replace(/^#/, '')
-  if (!/^[0-9a-f]{6}$/i.test(hex)) {
-    return `rgba(0, 0, 0, ${alpha})`
-  }
-  const r = Number.parseInt(hex.slice(0, 2), 16)
-  const g = Number.parseInt(hex.slice(2, 4), 16)
-  const b = Number.parseInt(hex.slice(4, 6), 16)
-  return `rgba(${r}, ${g}, ${b}, ${alpha})`
-}
-
-function parseUtcDate(value: string | null | undefined) {
-  if (!value) {
-    return null
-  }
-
-  const trimmed = value.trim()
-  if (!trimmed) {
-    return null
-  }
-
-  const normalized = /^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$/.test(trimmed)
-    ? `${trimmed.replace(' ', 'T')}Z`
-    : trimmed
-  const timestamp = Date.parse(normalized)
-  if (!Number.isFinite(timestamp)) {
-    return null
-  }
-
-  return timestamp
-}
-
-function resolveEventEndTimestamp(event: Event) {
-  const eventEnd = parseUtcDate(event.end_date)
-  const marketEnd = parseUtcDate(event.markets[0]?.end_time)
-
-  if (eventEnd != null && marketEnd != null) {
-    return Math.max(eventEnd, marketEnd)
-  }
-
-  if (eventEnd != null) {
-    return eventEnd
-  }
-
-  if (marketEnd != null) {
-    return marketEnd
-  }
-
-  return null
-}
-
-function inferIntervalMsFromSeriesSlug(seriesSlug: string | null | undefined) {
-  const normalized = seriesSlug?.trim().toLowerCase() ?? ''
-  if (!normalized) {
-    return 24 * 60 * 60 * 1000
-  }
-
-  if (normalized.includes('5m')) {
-    return 5 * 60 * 1000
-  }
-
-  if (normalized.includes('15m')) {
-    return 15 * 60 * 1000
-  }
-
-  if (normalized.includes('hourly') || normalized.includes('1h')) {
-    return 60 * 60 * 1000
-  }
-
-  if (normalized.includes('4h')) {
-    return 4 * 60 * 60 * 1000
-  }
-
-  return 24 * 60 * 60 * 1000
-}
-
-type CountdownUnit = 'day' | 'hr' | 'min' | 'sec'
-
-function countdownLabel(unit: CountdownUnit, value: number) {
-  if (unit === 'day') {
-    return value === 1 ? 'DAY' : 'DAYS'
-  }
-
-  const singular = unit.toUpperCase()
-  const plural = `${singular}S`
-  return value === 1 ? singular : plural
-}
-
-function toCountdownLeftLabel(showDays: boolean, days: number, hours: number, minutes: number, seconds: number) {
-  if (showDays) {
-    return `${days} ${days === 1 ? 'Day' : 'Days'} ${hours} ${hours === 1 ? 'Hr' : 'Hrs'} ${minutes} ${minutes === 1 ? 'Min' : 'Mins'}`
-  }
-
-  return `${hours} ${hours === 1 ? 'Hr' : 'Hrs'} ${minutes} ${minutes === 1 ? 'Min' : 'Mins'} ${seconds} ${seconds === 1 ? 'Sec' : 'Secs'}`
-}
-
-function getVisibleCountdownUnits(showDays: boolean, days: number, hours: number, minutes: number, seconds: number) {
-  if (showDays) {
-    return [
-      { unit: 'day' as const, value: days },
-      { unit: 'hr' as const, value: hours },
-      { unit: 'min' as const, value: minutes },
-    ]
-  }
-
-  return [
-    { unit: 'hr' as const, value: hours },
-    { unit: 'min' as const, value: minutes },
-    { unit: 'sec' as const, value: seconds },
-  ]
-}
-
-function formatDateAtTimezone(timestamp: number, timeZone: string) {
-  return new Intl.DateTimeFormat('en-US', {
-    month: 'short',
-    day: 'numeric',
-    year: 'numeric',
-    timeZone,
-  }).format(new Date(timestamp))
-}
-
-function formatTimeAtTimezone(timestamp: number, timeZone: string) {
-  return new Intl.DateTimeFormat('en-US', {
-    hour: 'numeric',
-    minute: '2-digit',
-    hour12: true,
-    timeZone,
-  }).format(new Date(timestamp))
-}
-
-function isUsEquityMarketOpen(timestamp: number) {
-  const parts = new Intl.DateTimeFormat('en-US', {
-    timeZone: 'America/New_York',
-    weekday: 'short',
-    hour: '2-digit',
-    minute: '2-digit',
-    hour12: false,
-  }).formatToParts(new Date(timestamp))
-
-  const weekday = parts.find(part => part.type === 'weekday')?.value ?? ''
-  const hourValue = Number(parts.find(part => part.type === 'hour')?.value ?? '0')
-  const minuteValue = Number(parts.find(part => part.type === 'minute')?.value ?? '0')
-  const minutesOfDay = hourValue * 60 + minuteValue
-
-  if (weekday === 'Sat' || weekday === 'Sun') {
-    return false
-  }
-
-  return minutesOfDay >= 9 * 60 + 30 && minutesOfDay < 16 * 60
-}
-
-function clampCountdownDigit(value: number) {
-  return Math.max(0, Math.min(9, value))
-}
-
-function RollingCountdownDigit({ digit }: { digit: number }) {
-  const nextDigit = clampCountdownDigit(digit)
-  const [currentDigit, setCurrentDigit] = useState(() => nextDigit)
-  const [previousDigit, setPreviousDigit] = useState<number | null>(null)
-  const [isAnimating, setIsAnimating] = useState(false)
-
-  if (nextDigit !== currentDigit) {
-    setPreviousDigit(currentDigit)
-    setCurrentDigit(nextDigit)
-    setIsAnimating(true)
-  }
-
-  function handleDigitTransitionEnd() {
-    if (!isAnimating) {
-      return
-    }
-
-    setIsAnimating(false)
-    setPreviousDigit(null)
-  }
-
-  return (
-    <span className="relative inline-flex h-[1em] w-[0.72em] overflow-hidden">
-      {previousDigit !== null && (
-        <span
-          className={cn(
-            'absolute inset-0 flex items-center justify-center transition-all duration-200 ease-out',
-            isAnimating ? '-translate-y-full opacity-0' : 'translate-y-0 opacity-100',
-          )}
-        >
-          {previousDigit}
-        </span>
-      )}
-      <span
-        className={cn(
-          'absolute inset-0 flex items-center justify-center transition-all duration-200 ease-out',
-          previousDigit === null
-            ? 'translate-y-0 opacity-100'
-            : isAnimating
-              ? 'translate-y-0 opacity-100'
-              : 'translate-y-full opacity-0',
-        )}
-        onTransitionEnd={handleDigitTransitionEnd}
-      >
-        {currentDigit}
-      </span>
-    </span>
-  )
-}
-
-function AnimatedCountdownValue({ value }: { value: number }) {
-  const safeValue = Math.max(0, Math.floor(value))
-  const digits = safeValue.toString().padStart(2, '0').split('')
-
-  return (
-    <span className="inline-flex items-center leading-none tabular-nums">
-      {digits.map((digit, index) => (
-        <RollingCountdownDigit
-          key={index}
-          digit={Number(digit)}
-        />
-      ))}
-    </span>
-  )
-}
 
 export function shouldUseLiveSeriesChart(event: Event, config: EventLiveChartConfig | null | undefined) {
   if (!config?.enabled) {
@@ -731,350 +114,48 @@ function EventLiveSeriesChartContent({
   config,
   subscriptionSymbol,
 }: EventLiveSeriesChartContentProps) {
-  const wsUrl = process.env.WS_LIVE_DATA_URL
   const site = useSiteIdentity()
   const { width: windowWidth } = useWindowSize()
   const liveColor = config.line_color || '#F59E0B'
-  const [nowMs, setNowMs] = useState(0)
-  const [data, setData] = useState<DataPoint[]>([])
-  const [persistedFallbackPrice, setPersistedFallbackPrice] = useState<PersistedLivePrice | null>(
-    () => readPersistedLivePrice(config.topic, subscriptionSymbol),
-  )
-  const [baselinePrice, setBaselinePrice] = useState<number | null>(null)
-  const [referenceSnapshot, setReferenceSnapshot] = useState<LiveSeriesPriceSnapshot | null>(null)
-  const [status, setStatus] = useState<'connecting' | 'live' | 'offline'>(
-    () => (wsUrl ? 'connecting' : 'offline'),
-  )
   const [activeView, setActiveView] = useState<'live' | 'market'>('live')
   const isLiveView = activeView === 'live'
   const startTimestamp = useMemo(() => parseUtcDate(event.start_date ?? null), [event.start_date])
   const explicitEndTimestamp = useMemo(() => resolveEventEndTimestamp(event), [event])
   const hasExplicitEndTimestamp = explicitEndTimestamp != null
+
+  const nowMs = useLiveSeriesClock(isLiveView)
   const endTimestamp = explicitEndTimestamp ?? nowMs
   const isEventClosed = hasExplicitEndTimestamp && nowMs >= endTimestamp
+
+  const {
+    referenceSnapshot,
+    baselinePrice,
+    setBaselinePrice,
+    persistedFallbackPrice: snapshotFallbackPrice,
+  } = useLiveSeriesPriceSnapshot({
+    config,
+    subscriptionSymbol,
+    explicitEndTimestamp,
+    startTimestamp,
+  })
+
+  const persistedFallbackPrice = snapshotFallbackPrice
+    ?? readPersistedLivePrice(config.topic, subscriptionSymbol)
+
+  const { data, status } = useLiveSeriesWebSocket({
+    topic: config.topic,
+    eventType: config.event_type,
+    subscriptionSymbol,
+    isLiveView,
+    setBaselinePrice,
+  })
+
   const isMarketClosed = useMemo(() => {
     if (config.topic.trim().toLowerCase() !== 'equity_prices') {
       return false
     }
     return !isUsEquityMarketOpen(nowMs)
   }, [config.topic, nowMs])
-
-  useEffect(() => {
-    if (!isLiveView) {
-      return
-    }
-
-    let frameId: number | null = null
-    let lastFrameTimestamp = 0
-
-    function animate(frameTimestamp: number) {
-      if (!document.hidden && frameTimestamp - lastFrameTimestamp >= LIVE_CLOCK_FRAME_MS) {
-        lastFrameTimestamp = frameTimestamp
-        setNowMs(Date.now())
-      }
-      frameId = window.requestAnimationFrame(animate)
-    }
-
-    frameId = window.requestAnimationFrame(animate)
-
-    function handleVisibilityChange() {
-      if (!document.hidden) {
-        setNowMs(Date.now())
-      }
-    }
-
-    document.addEventListener('visibilitychange', handleVisibilityChange)
-
-    return () => {
-      if (frameId != null) {
-        window.cancelAnimationFrame(frameId)
-      }
-      document.removeEventListener('visibilitychange', handleVisibilityChange)
-    }
-  }, [isLiveView])
-
-  useEffect(() => {
-    const seriesSlug = config.series_slug?.trim()
-    if (!seriesSlug) {
-      return
-    }
-
-    const snapshotEventEndMs = explicitEndTimestamp ?? Date.now()
-    if (!Number.isFinite(snapshotEventEndMs) || snapshotEventEndMs <= 0) {
-      return
-    }
-
-    const controller = new AbortController()
-    let isCancelled = false
-
-    async function loadPriceSnapshot() {
-      try {
-        const query = new URLSearchParams({
-          seriesSlug,
-          eventEndMs: String(snapshotEventEndMs),
-          activeWindowMinutes: String(config.active_window_minutes),
-        })
-        if (startTimestamp != null && startTimestamp > 0 && startTimestamp < snapshotEventEndMs) {
-          query.set('eventStartMs', String(startTimestamp))
-        }
-
-        const response = await fetch(`/api/price-reference/live-series?${query.toString()}`, {
-          cache: 'no-store',
-          signal: controller.signal,
-        })
-
-        if (!response.ok) {
-          return
-        }
-
-        const payload = await response.json() as LiveSeriesPriceSnapshot
-        if (isCancelled) {
-          return
-        }
-
-        setReferenceSnapshot(payload)
-
-        if (typeof payload.opening_price === 'number' && Number.isFinite(payload.opening_price) && payload.opening_price > 0) {
-          setBaselinePrice(payload.opening_price)
-        }
-
-        const fallbackPrice = normalizeLiveChartPrice(
-          payload.latest_price ?? payload.closing_price ?? Number.NaN,
-          config.topic,
-        )
-
-        if (typeof fallbackPrice === 'number') {
-          const rawFallbackTimestamp = payload.latest_source_timestamp_ms ?? payload.event_window_end_ms ?? Date.now()
-          const minTimestamp = Date.now() - LIVE_DATA_RETENTION_MS + 1000
-          const fallbackTimestamp = Math.max(rawFallbackTimestamp, minTimestamp)
-          writePersistedLivePrice(config.topic, subscriptionSymbol, fallbackPrice, fallbackTimestamp)
-          setPersistedFallbackPrice({
-            price: fallbackPrice,
-            timestamp: fallbackTimestamp,
-          })
-        }
-      }
-      catch {
-      }
-    }
-
-    loadPriceSnapshot()
-
-    return () => {
-      isCancelled = true
-      controller.abort()
-    }
-  }, [config.active_window_minutes, config.series_slug, config.topic, explicitEndTimestamp, startTimestamp, subscriptionSymbol])
-
-  useEffect(() => {
-    if (!isLiveView) {
-      return
-    }
-
-    if (!wsUrl) {
-      return
-    }
-    // Intentionally keep WS active regardless of event close to preserve always-live behavior.
-    const resolvedWsUrl = wsUrl
-
-    let isActive = true
-    let ws: WebSocket | null = null
-
-    function buildSubscriptionPayload(action: 'subscribe' | 'unsubscribe') {
-      const filters = JSON.stringify({
-        symbol: subscriptionSymbol,
-      })
-
-      return JSON.stringify({
-        action,
-        subscriptions: [
-          {
-            topic: config.topic,
-            type: config.event_type,
-            filters,
-          },
-        ],
-      })
-    }
-
-    function handleOpen() {
-      if (!ws) {
-        return
-      }
-      setStatus('connecting')
-      ws.send(buildSubscriptionPayload('subscribe'))
-    }
-
-    function handleMessage(eventMessage: MessageEvent<string>) {
-      if (!isActive) {
-        return
-      }
-
-      let payload: any
-      try {
-        payload = JSON.parse(eventMessage.data)
-      }
-      catch {
-        return
-      }
-
-      const updates = extractLivePriceUpdates(payload, config.topic, subscriptionSymbol, Date.now())
-      const normalizedUpdates = updates
-        .map((update) => {
-          const normalizedPrice = normalizeLiveChartPrice(update.price, config.topic)
-          if (normalizedPrice == null) {
-            return null
-          }
-
-          return {
-            ...update,
-            price: normalizedPrice,
-          }
-        })
-        .filter((update): update is { price: number, timestamp: number, symbol: string | null } => update !== null)
-
-      const messageIsSnapshot = isSnapshotMessage(payload)
-      const messageHasBatchUpdates = normalizedUpdates.length > 1
-      const wsUpdatesForRender = LIVE_WS_USE_ONLY_LAST_UPDATE_PER_MESSAGE
-        ? (messageIsSnapshot || messageHasBatchUpdates ? normalizedUpdates : normalizedUpdates.slice(-1))
-        : normalizedUpdates
-
-      if (!wsUpdatesForRender.length) {
-        return
-      }
-
-      setStatus('live')
-      const latest = wsUpdatesForRender.at(-1)
-      if (latest) {
-        writePersistedLivePrice(config.topic, subscriptionSymbol, latest.price, latest.timestamp)
-      }
-
-      setData((prev) => {
-        const arrivalTimestamp = Date.now()
-        const cutoff = arrivalTimestamp - LIVE_DATA_RETENTION_MS
-
-        if (messageIsSnapshot && wsUpdatesForRender.length > 1) {
-          let lastSnapshotTimestamp: number | null = null
-          const snapshotPoints: DataPoint[] = []
-
-          for (const update of wsUpdatesForRender) {
-            let pointTimestamp = update.timestamp
-            if (!Number.isFinite(pointTimestamp)) {
-              continue
-            }
-
-            pointTimestamp = Math.max(cutoff + 1, Math.min(pointTimestamp, arrivalTimestamp))
-            if (lastSnapshotTimestamp !== null && pointTimestamp <= lastSnapshotTimestamp) {
-              pointTimestamp = lastSnapshotTimestamp + 1
-            }
-
-            snapshotPoints.push({
-              date: new Date(pointTimestamp),
-              [SERIES_KEY]: update.price,
-            })
-            lastSnapshotTimestamp = pointTimestamp
-          }
-
-          if (snapshotPoints.length > 0) {
-            return snapshotPoints.slice(-MAX_POINTS)
-          }
-        }
-
-        let next = keepWithinLiveWindow(prev, cutoff)
-        const lastPoint = next.length ? next.at(-1) : null
-        let lastTimestamp = lastPoint ? lastPoint.date.getTime() : null
-
-        for (const update of wsUpdatesForRender) {
-          // Anchor incoming points to arrival time to avoid delayed-source timestamp jumps.
-          let pointTimestamp = Math.max(update.timestamp, arrivalTimestamp)
-
-          if (lastTimestamp !== null && pointTimestamp <= lastTimestamp) {
-            pointTimestamp = lastTimestamp + 1
-          }
-
-          const nextPoint: DataPoint = {
-            date: new Date(pointTimestamp),
-            [SERIES_KEY]: update.price,
-          }
-
-          next = [...next, nextPoint].slice(-MAX_POINTS)
-          lastTimestamp = pointTimestamp
-        }
-
-        return next
-      })
-
-      setBaselinePrice(current => current ?? wsUpdatesForRender[0]?.price ?? null)
-    }
-
-    function handleError() {
-      if (isActive) {
-        setStatus('offline')
-      }
-    }
-
-    let reconnectController: ReturnType<typeof createWebSocketReconnectController> | null = null
-
-    function clearReconnect() {
-      reconnectController?.clearReconnect()
-    }
-
-    function handleVisibilityChange() {
-      reconnectController?.handleVisibilityChange()
-    }
-
-    function scheduleReconnect() {
-      reconnectController?.scheduleReconnect()
-    }
-
-    function handleClose() {
-      if (!isActive) {
-        return
-      }
-      setStatus('offline')
-      scheduleReconnect()
-    }
-
-    function connect() {
-      if (!isActive || ws || document.hidden) {
-        return
-      }
-      ws = new WebSocket(resolvedWsUrl)
-      ws.addEventListener('open', handleOpen)
-      ws.addEventListener('message', handleMessage)
-      ws.addEventListener('error', handleError)
-      ws.addEventListener('close', handleClose)
-    }
-
-    reconnectController = createWebSocketReconnectController({
-      connect,
-      getWebSocket: () => ws,
-      isActive: () => isActive,
-      resetWebSocket: () => {
-        ws = null
-      },
-    })
-
-    connect()
-    document.addEventListener('visibilitychange', handleVisibilityChange)
-
-    return () => {
-      isActive = false
-      setStatus('offline')
-      clearReconnect()
-      document.removeEventListener('visibilitychange', handleVisibilityChange)
-      if (ws) {
-        if (ws.readyState === WebSocket.OPEN) {
-          ws.send(buildSubscriptionPayload('unsubscribe'))
-        }
-        ws.removeEventListener('open', handleOpen)
-        ws.removeEventListener('message', handleMessage)
-        ws.removeEventListener('error', handleError)
-        ws.removeEventListener('close', handleClose)
-        ws.close()
-      }
-    }
-  }, [config.event_type, config.topic, isLiveView, wsUrl, subscriptionSymbol])
 
   const series = useMemo<SeriesConfig[]>(
     () => ([{
@@ -1127,6 +208,7 @@ function EventLiveSeriesChartContent({
 
     return inferIntervalMsFromSeriesSlug(config.series_slug)
   }, [config.active_window_minutes, config.series_slug, referenceSnapshot?.interval_ms])
+
   const tradingWindowStartMs = useMemo(() => {
     if (startTimestamp != null && startTimestamp > 0 && startTimestamp < endTimestamp) {
       return startTimestamp
@@ -1139,6 +221,7 @@ function EventLiveSeriesChartContent({
 
     return endTimestamp - tradingWindowMs
   }, [endTimestamp, referenceSnapshot?.event_window_start_ms, startTimestamp, tradingWindowMs])
+
   const isTradingWindowActive = !isEventClosed && nowMs >= tradingWindowStartMs
 
   const renderData = useMemo(() => {
@@ -1230,6 +313,7 @@ function EventLiveSeriesChartContent({
 
     return buildAxis(values, priceDisplayDigits)
   })()
+
   const currentLineTop = (() => {
     if (currentPrice == null) {
       return null
@@ -1242,6 +326,7 @@ function EventLiveSeriesChartContent({
     const clamped = Math.max(0, Math.min(1, ratio))
     return marginTop + innerHeight - innerHeight * clamped
   })()
+
   const targetLine = (() => {
     if (!isTradingWindowActive || resolvedBaselinePrice == null || !Number.isFinite(resolvedBaselinePrice)) {
       return null
@@ -1262,9 +347,11 @@ function EventLiveSeriesChartContent({
       isBelow: ratio < 0,
     }
   })()
+
   const targetLineGuideColor = hexToRgba('#94a3b8', 0.62)
   const targetBadgeColor = '#94a3b8'
   const currentPriceGuideColor = hexToRgba(liveColor, 0.62)
+
   const countdown = useMemo(() => {
     const totalSeconds = Math.max(0, Math.floor((endTimestamp - nowMs) / 1000))
     const showDays = totalSeconds > 24 * 60 * 60
@@ -1284,7 +371,9 @@ function EventLiveSeriesChartContent({
       seconds,
     }
   }, [endTimestamp, nowMs])
+
   const shouldShowCountdown = hasExplicitEndTimestamp && !isEventClosed && countdown.totalSeconds > 0
+
   const xAxisTickValues = useMemo(() => {
     const startMs = nowMs - LIVE_WINDOW_MS
     const visibleStartMs = startMs + LIVE_X_AXIS_LEFT_LABEL_GUARD_MS
@@ -1306,6 +395,7 @@ function EventLiveSeriesChartContent({
       new Date(nowMs),
     ]
   }, [nowMs])
+
   const liveXAxisDomain = useMemo(
     () => ({
       start: new Date(nowMs - LIVE_WINDOW_MS),
@@ -1313,6 +403,7 @@ function EventLiveSeriesChartContent({
     }),
     [nowMs],
   )
+
   const visibleCountdownUnits = useMemo(
     () => getVisibleCountdownUnits(
       countdown.showDays,
@@ -1323,6 +414,7 @@ function EventLiveSeriesChartContent({
     ),
     [countdown.showDays, countdown.days, countdown.hours, countdown.minutes, countdown.seconds],
   )
+
   const countdownLeftLabel = useMemo(
     () => toCountdownLeftLabel(
       countdown.showDays,
@@ -1333,6 +425,7 @@ function EventLiveSeriesChartContent({
     ),
     [countdown.showDays, countdown.days, countdown.hours, countdown.minutes, countdown.seconds],
   )
+
   const etDateLabel = useMemo(
     () => formatDateAtTimezone(endTimestamp, 'America/New_York'),
     [endTimestamp],
@@ -1349,21 +442,6 @@ function EventLiveSeriesChartContent({
     () => formatTimeAtTimezone(endTimestamp, 'UTC'),
     [endTimestamp],
   )
-  const isMarketView = activeView === 'market'
-  const isLiveChartView = activeView === 'live'
-  const liveSwitchIconStyle = isLiveChartView
-    ? {
-        color: liveColor,
-      }
-    : undefined
-  const switchThumbStyle = isLiveChartView
-    ? {
-        transform: 'translateX(2rem)',
-        backgroundColor: hexToRgba(liveColor, 0.2),
-      }
-    : {
-        transform: 'translateX(0)',
-      }
 
   const watermark = useMemo(
     () => ({
@@ -1373,288 +451,40 @@ function EventLiveSeriesChartContent({
     }),
     [site.logoImageUrl, site.logoSvg, site.name],
   )
-  const countdownEndedLogo = (watermark.iconSvg || watermark.label)
-    ? (
-        <div
-          className="pointer-events-none flex items-center gap-1 text-xl text-muted-foreground opacity-50 select-none"
-          aria-hidden
-        >
-          {watermark.iconSvg
-            ? (
-                <SiteLogoIcon
-                  logoSvg={watermark.iconSvg}
-                  logoImageUrl={watermark.iconImageUrl}
-                  alt={`${watermark.label} logo`}
-                  className="size-[1em] **:fill-current **:stroke-current"
-                  imageClassName="size-[1em] object-contain"
-                  size={20}
-                />
-              )
-            : null}
-          {watermark.label
-            ? (
-                <span className="font-semibold">
-                  {watermark.label}
-                </span>
-              )
-            : null}
-        </div>
-      )
-    : null
-
-  const viewSwitch = (
-    <div className="relative z-0 flex items-center rounded-lg border border-border bg-background/70 p-0.5">
-      <span
-        className={cn(
-          'pointer-events-none absolute top-0.5 left-0.5 z-0 size-8 rounded-md transition-all duration-300 ease-out',
-          !isLiveChartView && 'bg-primary/30',
-        )}
-        style={switchThumbStyle}
-      />
-      <button
-        type="button"
-        onClick={() => setActiveView('market')}
-        className={cn(
-          'relative z-1 flex size-8 items-center justify-center rounded-md transition-colors',
-          isMarketView
-            ? 'text-primary'
-            : 'bg-transparent text-muted-foreground hover:bg-muted',
-        )}
-        aria-label="Show market chart"
-      >
-        <ChartLineIcon className="size-[18px]" />
-      </button>
-      <button
-        type="button"
-        onClick={() => setActiveView('live')}
-        className={cn(
-          'relative z-1 flex size-8 items-center justify-center rounded-md transition-colors',
-          !isLiveChartView && 'bg-transparent text-muted-foreground hover:bg-muted',
-        )}
-        style={liveSwitchIconStyle}
-        aria-label="Show live chart"
-      >
-        {config.icon_path
-          ? (
-              <span
-                className="block size-4 bg-current"
-                aria-hidden
-                style={{
-                  WebkitMaskImage: `url(${config.icon_path})`,
-                  maskImage: `url(${config.icon_path})`,
-                  WebkitMaskPosition: 'center',
-                  maskPosition: 'center',
-                  WebkitMaskRepeat: 'no-repeat',
-                  maskRepeat: 'no-repeat',
-                  WebkitMaskSize: 'contain',
-                  maskSize: 'contain',
-                }}
-              />
-            )
-          : <span className="size-2 rounded-full bg-current" />}
-      </button>
-    </div>
-  )
 
   return (
     <div className="grid gap-4">
       {isLiveView
         ? (
             <div className="grid gap-1">
-              <div className="flex flex-wrap items-end gap-4 pr-4 pl-0 sm:pr-6 sm:pl-0">
-                <div className="flex flex-wrap items-end gap-5">
-                  <div>
-                    <div className="text-[11px] font-semibold tracking-[0.12em] text-muted-foreground uppercase">
-                      Price To Beat
-                    </div>
-                    <div className="mt-1 text-[22px] leading-none font-semibold text-muted-foreground tabular-nums">
-                      {resolvedBaselinePrice != null ? formatUsd(resolvedBaselinePrice, headerPriceDisplayDigits) : '--'}
-                    </div>
-                  </div>
-                  <div className="hidden h-10 w-px bg-border sm:block" />
-                  <div>
-                    <div
-                      className="flex items-center gap-2 text-[11px] font-semibold tracking-[0.12em] uppercase"
-                      style={{ color: liveColor }}
-                    >
-                      <span>Current Price</span>
-                      {delta != null && (
-                        <span className={`inline-flex items-center gap-0.5 text-[11px] font-semibold ${delta >= 0
-                          ? 'text-yes'
-                          : 'text-no'}`}
-                        >
-                          <TriangleIcon
-                            className={`size-2.5 ${delta >= 0 ? '' : 'rotate-180'}`}
-                            fill="currentColor"
-                            stroke="none"
-                          />
-                          {formatUsd(Math.abs(delta), deltaDisplayDigits)}
-                        </span>
-                      )}
-                    </div>
-                    <div
-                      className="mt-1 text-[22px] leading-none font-semibold tabular-nums"
-                      style={{ color: liveColor }}
-                    >
-                      {currentPrice != null ? formatUsd(currentPrice, headerPriceDisplayDigits) : '--'}
-                    </div>
-                  </div>
-                </div>
-                {shouldShowCountdown
-                  ? (
-                      <Tooltip>
-                        <TooltipTrigger asChild>
-                          <button
-                            type="button"
-                            className="mr-[-4px] ml-auto grid justify-items-end gap-1 text-left sm:mr-[-6px]"
-                          >
-                            <div className="flex items-end gap-3">
-                              {visibleCountdownUnits.map(({ unit, value }) => (
-                                <div key={unit} className="min-w-11 text-right">
-                                  <div
-                                    className={cn(
-                                      'text-[22px] leading-none font-semibold tabular-nums',
-                                      isTradingWindowActive ? 'text-red-500' : 'text-muted-foreground',
-                                    )}
-                                  >
-                                    <AnimatedCountdownValue value={value} />
-                                  </div>
-                                  <div
-                                    className="
-                                      mt-1 text-2xs font-semibold tracking-[0.08em] text-muted-foreground uppercase
-                                    "
-                                  >
-                                    {countdownLabel(unit, value)}
-                                  </div>
-                                </div>
-                              ))}
-                            </div>
-                            <span className="sr-only">{status}</span>
-                          </button>
-                        </TooltipTrigger>
-                        <TooltipContent align="end" className="w-72 rounded-xl p-3 text-left">
-                          <div className="grid gap-2.5">
-                            <div className="flex items-center justify-between gap-3">
-                              <div className="inline-flex items-center gap-2 text-red-500">
-                                <span className="relative inline-flex size-2.5 items-center justify-center">
-                                  <span
-                                    className="
-                                      absolute inset-0 m-auto inline-flex size-2.5 animate-ping rounded-full
-                                      bg-red-500/45
-                                    "
-                                  />
-                                  <span
-                                    className="relative inline-flex size-2 rounded-full bg-red-500"
-                                  />
-                                </span>
-                                <span className="text-xs font-semibold tracking-[0.08em] uppercase">Live</span>
-                              </div>
-                              <div className="text-sm">
-                                <span className="font-semibold text-foreground">{countdownLeftLabel}</span>
-                                <span className="ml-1 text-muted-foreground">left</span>
-                              </div>
-                            </div>
+              <EventLiveSeriesChartHeader
+                resolvedBaselinePrice={resolvedBaselinePrice}
+                headerPriceDisplayDigits={headerPriceDisplayDigits}
+                currentPrice={currentPrice}
+                delta={delta}
+                deltaDisplayDigits={deltaDisplayDigits}
+                liveColor={liveColor}
+                shouldShowCountdown={shouldShowCountdown}
+                isEventClosed={isEventClosed}
+                isTradingWindowActive={isTradingWindowActive}
+                visibleCountdownUnits={visibleCountdownUnits}
+                countdownLeftLabel={countdownLeftLabel}
+                etDateLabel={etDateLabel}
+                etTimeLabel={etTimeLabel}
+                utcDateLabel={utcDateLabel}
+                utcTimeLabel={utcTimeLabel}
+                status={status}
+                watermark={watermark}
+              />
 
-                            <div className="text-xs text-muted-foreground">Resolution time</div>
-
-                            <div className="grid gap-2 text-sm text-foreground">
-                              <div className="flex items-center gap-2">
-                                <span
-                                  className="
-                                    inline-flex h-6 min-w-9 items-center justify-center rounded-md bg-muted px-2 text-xs
-                                    font-semibold
-                                  "
-                                >
-                                  ET
-                                </span>
-                                <span className="tabular-nums">{etDateLabel}</span>
-                                <span className="ml-auto tabular-nums">{etTimeLabel}</span>
-                              </div>
-                              <div className="flex items-center gap-2">
-                                <span
-                                  className="
-                                    inline-flex h-6 min-w-9 items-center justify-center rounded-md bg-muted px-2 text-xs
-                                    font-semibold
-                                  "
-                                >
-                                  UTC
-                                </span>
-                                <span className="tabular-nums">{utcDateLabel}</span>
-                                <span className="ml-auto tabular-nums">{utcTimeLabel}</span>
-                              </div>
-                            </div>
-                          </div>
-                        </TooltipContent>
-                      </Tooltip>
-                    )
-                  : isEventClosed
-                    ? (
-                        <div className="mr-[-4px] ml-auto sm:mr-[-6px]">
-                          {countdownEndedLogo}
-                        </div>
-                      )
-                    : null}
-              </div>
-
-              <div className="relative z-0 pr-4 pl-0 sm:pr-6 sm:pl-0">
-                {targetLine && (
-                  <>
-                    <div
-                      className="pointer-events-none absolute right-4 left-0 z-1 h-px sm:right-6"
-                      style={{
-                        top: `${targetLine.badgeTop}px`,
-                        backgroundImage: `repeating-linear-gradient(
-                          to right,
-                          ${targetLineGuideColor} 0px,
-                          ${targetLineGuideColor} 8px,
-                          transparent 8px,
-                          transparent 14px
-                        )`,
-                      }}
-                    />
-                    <span
-                      className="
-                        pointer-events-none absolute right-4 z-1 inline-flex -translate-y-1/2 items-center
-                        sm:right-6
-                      "
-                      style={{ top: `${targetLine.badgeTop}px` }}
-                    >
-                      <span
-                        aria-hidden
-                        className="
-                          relative z-0 -mr-px inline-block h-3.5 w-2 [clip-path:polygon(0_50%,100%_0,100%_100%)]
-                        "
-                        style={{ backgroundColor: targetBadgeColor }}
-                      />
-                      <span
-                        className="
-                          relative z-1 inline-flex items-center gap-0.5 rounded-[4px] px-2 py-1 pl-2 text-xs
-                          font-semibold text-white
-                        "
-                        style={{ backgroundColor: targetBadgeColor }}
-                      >
-                        <span>Target</span>
-                        {targetLine.isAbove && <ChevronsUpIcon className="size-3.5 animate-pulse" />}
-                        {targetLine.isBelow && <ChevronsDownIcon className="size-3.5 animate-pulse" />}
-                      </span>
-                    </span>
-                  </>
-                )}
-                {currentLineTop != null && (
-                  <div
-                    className="pointer-events-none absolute right-4 left-0 z-2 mr-2 h-px sm:right-6"
-                    style={{
-                      top: `${currentLineTop}px`,
-                      backgroundImage: `repeating-linear-gradient(
-                        to right,
-                        ${currentPriceGuideColor} 0px,
-                        ${currentPriceGuideColor} 8px,
-                        transparent 8px,
-                        transparent 14px
-                      )`,
-                    }}
-                  />
-                )}
+              <div className={cn('relative z-0 pr-4 pl-0 sm:pr-6 sm:pl-0')}>
+                <EventLiveSeriesChartOverlay
+                  targetLine={targetLine}
+                  targetLineGuideColor={targetLineGuideColor}
+                  targetBadgeColor={targetBadgeColor}
+                  currentLineTop={currentLineTop}
+                  currentPriceGuideColor={currentPriceGuideColor}
+                />
                 <PredictionChart
                   data={renderData}
                   series={series}
@@ -1738,7 +568,14 @@ function EventLiveSeriesChartContent({
         currentEventSlug={event.slug}
         seriesEvents={seriesEvents}
         variant="live"
-        rightSlot={viewSwitch}
+        rightSlot={(
+          <EventLiveSeriesViewSwitch
+            activeView={activeView}
+            setActiveView={setActiveView}
+            liveColor={liveColor}
+            config={config}
+          />
+        )}
       />
     </div>
   )
